Oh Crud
Abstracted the database code into standard crud methods (with one and many variants for each). The return objects are also abstracted. This is all stuffed into a database package, with a bindings file for pymongo. Additional bindings for other databases can be created by writing a new set of bindings to match the current ones and changing the import in the pacakge __init__.py file.