BoM Parse and Store
Parsing, DB lookup and storing of BoM's now works.

Got rid of the modules object, and now we pass down a ref to the top level app. Main class functional objects are stored in the app directly so can be accessed like manf.factory instead of manf.modules.factory.

Also some tidy ups and that.